+++ internal/engine/player.go
@@ -4,6 +4,6 @@ type (
 	// Player represents a player of the mancala game
 	Player struct {
 		Name  string `json:"name"`
-		Score int    `json:"score"`
+		Score int64  `json:"score"`
 	}
 )
